+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/mediacodec/AsynchronousMediaCodecAdapter.java
@@ -275,6 +275,12 @@ import java.nio.ByteBuffer;
   }
 
   @Override
+  public boolean registerOnBufferAvailableListener(OnBufferAvailableListener listener) {
+    asynchronousMediaCodecCallback.setOnBufferAvailableListener(listener);
+    return true;
+  }
+
+  @Override
   public void setOutputSurface(Surface surface) {
     codec.setOutputSurface(surface);
   }